last update

+++ asm/asmlib/src/ft_puthex_fd.c
@@ -6,13 +6,13 @@
 /*   By: yhetman <yhetman@student.unit.ua>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/24 13:07:11 by yhetman           #+#    #+#             */
-/*   Updated: 2019/10/24 13:11:57 by yhetman          ###   ########.fr       */
+/*   Updated: 2019/10/24 18:13:11 by blukasho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/asmlib.h"
 
-void ft_puthex_fd(int hex, int fd)
+void ft_puthex_fd(unsigned long long hex, int fd)
 {
 	if (hex < 256)
 	{
